requirements.txt
altgraph==0.17
colorama==0.4.4
colorlog==4.7.2
et-xmlfile==1.0.1
future==0.18.2
jdcal==1.4.1
numpy==1.19.5
openpyxl==3.0.6
pandas==1.2.1
pefile==2019.4.18
Pillow==8.0.1
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
python-dateutil==2.8.1
pytz==2020.5
pywin32-ctypes==0.2.0
six==1.15.0
xlrd==2.0.1
